+++ test/hop_tasks/_hop_tasks.dart
@@ -1,6 +1,8 @@
 library test_hop_tasks;
 
 import 'dart:io';
+import 'package:bot/bot.dart';
+import 'package:bot/hop.dart';
 import 'package:bot/hop_tasks.dart';
 import 'package:unittest/unittest.dart';
 import '../hop/_hop.dart';